new file

+++ add_2_numbers.cpp
@@ -3,10 +3,10 @@ using namespace std;
 
 int main(){
     cout<<"Enter Numbers: ";
-    int numb1, numb2;
+    long long int numb1, numb2;
     cin>>numb1>>numb2;
     
-    int sum = numb1 + numb2;
+    long long int sum = numb1 + numb2;
     cout<<"The sum: "<< sum<< endl;
     return 0;
 }
